+++ src-tauri/src/lib.rs
@@ -1,6 +1,13 @@
+mod save;
+
 #[cfg_attr(mobile, tauri::mobile_entry_point)]
 pub fn run() {
   tauri::Builder::default()
+    .invoke_handler(tauri::generate_handler![
+      save::save_load,
+      save::save_store,
+      save::save_location
+    ])
     .setup(|app| {
       if cfg!(debug_assertions) {
         app.handle().plugin(